Feature/7345 fix mail working in production

+++ tear-down-local.sh
@@ -1,3 +1,9 @@
 #!/bin/bash
 
+context=`kubectl config current-context`
+if [[ "$context" != "docker-desktop" ]]; then 
+  >&2 echo "Not Local Context - $context"
+  exit 1
+fi
+
 helm delete bc -n buyingcatalogue
